Rename module IdentifierGenerator to PersistentIdentifier
