+++ Kitodo-PersistentIdentifier/pom.xml
@@ -1,10 +1,22 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+ *
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ *
+-->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.kitodo.identifiergenerator</groupId>
-    <artifactId>kitodo-identifiergenerator</artifactId>
+    <groupId>org.kitodo.persistentidentifier</groupId>
+    <artifactId>kitodo-persistentidentifier</artifactId>
     <parent>
         <artifactId>kitodo-production</artifactId>
         <groupId>org.kitodo</groupId>
@@ -14,7 +26,7 @@
         <name>kitodo.org</name>
         <url>http://kitodo.org</url>
     </organization>
-    <name>Kitodo - Identifier Generator</name>
+    <name>Kitodo - Persistent Identifier</name>
     <url>https://github.com/kitodo/kitodo-production</url>
 
     <properties>
