+++ Kitodo-PersistentIdentifier/src/test/java/org/kitodo/persistentidentifier/DNBURNIdentifierGeneratorTest.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package org.kitodo.identifiergenerator;
+package org.kitodo.persistentidentifier;
 
 import org.junit.Test;
