+++ pom.xml
@@ -89,7 +89,7 @@
         <module>Kitodo-DataManagement</module>
         <module>Kitodo-API</module>
         <module>Kitodo-FileManagement</module>
-        <module>Kitodo-IdentifierGenerator</module>
+        <module>Kitodo-PersistentIdentifier</module>
         <module>Kitodo</module>
     </modules>
